+++ src/DefaultScanType/ImageNotRef.php
@@ -3,28 +3,26 @@
  * Author: Tony Chen
  */
 
-namespace XcxProfiler\DefScanType;
+namespace XcxProfiler\DefaultScanType;
 
 use XcxProfiler\File;
 use XcxProfiler\Filter\Filter;
 use XcxProfiler\Filter\FilterIsSpecifiedExtensionFile;
-use XcxProfiler\Filter\FilterNotUploadToRemote;
-use XcxProfiler\Filter\FilterSizeGreaterThan;
+use XcxProfiler\Filter\FilterNotRef;
 use XcxProfiler\ScanDir;
 
-class ImageNotUploadToRemote extends DefScanBase {
+class ImageNotRef extends DefScanBase {
     /**
      * 获取过滤器列表
      * @return Filter
      */
-    public function getFilerChain(): Filter {
-        #过滤器
+    protected function getFilerChain(): Filter {
+        // 过滤器
         $filter = new FilterIsSpecifiedExtensionFile(
             FilterIsSpecifiedExtensionFile::FILE_TYPE_PNG |
             FilterIsSpecifiedExtensionFile::FILE_TYPE_JPG
         );
-        $filter->setNextFilter(new FilterSizeGreaterThan(21))
-               ->setNextFilter(new FilterNotUploadToRemote($this->getScanFullPath()));
+        $filter->setNextFilter(new FilterNotRef($this->getScanFullPath()));
         return $filter;
     }
 
@@ -32,7 +30,7 @@ class ImageNotUploadToRemote extends DefScanBase {
      * 获取要被过滤的文件列表
      * @return File[]
      */
-    public function getCheckFiles(): array {
+    protected function getCheckFiles(): array {
         return ScanDir::getInstance($this->getScanFullPath())->getAllFile();
     }
 }
