+++ src/DefaultScanType/ImageNotUploadToRemote.php
@@ -3,26 +3,28 @@
  * Author: Tony Chen
  */
 
-namespace XcxProfiler\DefScanType;
+namespace XcxProfiler\DefaultScanType;
 
 use XcxProfiler\File;
 use XcxProfiler\Filter\Filter;
 use XcxProfiler\Filter\FilterIsSpecifiedExtensionFile;
-use XcxProfiler\Filter\FilterNotRef;
+use XcxProfiler\Filter\FilterNotUploadToRemote;
+use XcxProfiler\Filter\FilterSizeGreaterThan;
 use XcxProfiler\ScanDir;
 
-class ImageNotRef extends DefScanBase {
+class ImageNotUploadToRemote extends DefScanBase {
     /**
      * 获取过滤器列表
      * @return Filter
      */
-    public function getFilerChain(): Filter {
-        // 过滤器
+    protected function getFilerChain(): Filter {
+        #过滤器
         $filter = new FilterIsSpecifiedExtensionFile(
             FilterIsSpecifiedExtensionFile::FILE_TYPE_PNG |
             FilterIsSpecifiedExtensionFile::FILE_TYPE_JPG
         );
-        $filter->setNextFilter(new FilterNotRef($this->getScanFullPath()));
+        $filter->setNextFilter(new FilterSizeGreaterThan(21))
+               ->setNextFilter(new FilterNotUploadToRemote($this->getScanFullPath()));
         return $filter;
     }
 
@@ -30,7 +32,7 @@ class ImageNotRef extends DefScanBase {
      * 获取要被过滤的文件列表
      * @return File[]
      */
-    public function getCheckFiles(): array {
+    protected function getCheckFiles(): array {
         return ScanDir::getInstance($this->getScanFullPath())->getAllFile();
     }
 }
